Added dummy cache.
From now on mass keeps all the AWS data cached in /tmp directory to avoid
querying AWS API with every run. This step is necessary for users who have
more 50+ servers as mass:

- queries AWS for all the instances data
- queries AWS for all tags used by discovered instances

It takes a while in real life environment but thanks to caching it's just
the first run user needs to survive.